Workbooks and alert templates for AODS customers.
AODS is an Azure-connected edge platform deployed in the customer premise which in this case is a highly secure telco operator edge environment in the telco network, enabling 5G Core deployments.

One of the intents is to show Telco operators the value add that this cloud offering brings along and onboard them to this platform as quickly as possible. Workbooks and Alerts have been identified as one of the starting points and these are based of telemetry from multiple layers in stack.