Update Warm and Cold Form Load Data by Table.kql
Fix: 
Changed join from inner to fullouter. This was done to fix where the query was missing table data when there was not both cold and warm loads for a table.

Enhanced:
Added a coalesce in the project to unify the table data. Also added the option to allow a list of tables and also added sync request time to the results.